Settings: Fix setting device name based on market name
The current code for setting device name based on market name is incorrect; this means that if the required property exists on a device, it will always use the market name regardless if the user has customized their device name.

Fix this by checking if the market name prop exists if and only if there is no customized device name, then falling back to the device model if said prop doesn't exist.

Signed-off-by: John Vincent <git@tenseventyseven.cf>